% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{bp_gap.model}
\alias{bp_gap.model}
\title{Time series model for the best-practice gap}
\usage{
bp_gap.model(data, benchmark, arima.order, drift)
}
\description{
Time series model for the best-practice gap
}
\keyword{internal}
