% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_predict.R
\name{compute_CI}
\alias{compute_CI}
\title{Function to generate correlated prediction intervals from a mvrnorm}
\usage{
compute_CI(pred_results, m1, m2, m3, h, iter, ci, cou)
}
\description{
Function to generate correlated prediction intervals from a mvrnorm
}
\keyword{internal}
