% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{DoubleGap}
\alias{DoubleGap}
\title{Fit the DoubleGap Life Expectancy Forecasting Model}
\usage{
DoubleGap(DF, DM, age, country, years, arima.order = c(0, 1, 0),
  drift = FALSE, tau = NULL, A = NULL)
}
\arguments{
\item{DF}{data.frame containing life expectancy records for females. 
The table must contain the following 4 columns: country, year, age, ex.}

\item{DM}{data.frame containing life expectancy records for males. 
The table must have the same format and dimensions as \code{DF}.}

\item{age}{Indicate the age for which the model to be fitted. 
Assuming \code{DF} and \code{DM} contain recods for different ages, this 
argument it is used to subset the data. If you want to fit the model for age 0, 
add age = 0. Type: scalar.}

\item{country}{Indicate for which contry you want to fit the model. The country
name or code must exist in \code{DF} and \code{DM}.
Type: character.}

\item{years}{Period of time to be used. Type: numeric vector.}

\item{arima.order}{A specification of the the ARIMA model to be used in 
fitting the best-practice gap. The ARIMA order is country specific.
The three integer components (p, d, q) are the AR order, 
the degree of differencing, and the MA order.}

\item{drift}{Indicate whether the ARIMA model should include a linear drift 
term or not. Type: logical value.}

\item{tau}{The level of female life expectancy at which the sex-gap is 
expected to stop widening and to start narrowing. If \code{NULL} 
then the model will run an algorithm to find it.}

\item{A}{The level of female life expectancy where we assume no further 
change in the sex-gap. If \code{NULL} the model will estimate it.}
}
\value{
The output is of \code{"DoubleGap"} class with the components:

\item{input}{List with arguments provided in input. Saved for convenience.}

\item{call}{Short information about the model.}

\item{coefficients}{Estimated coefficients.}

\item{fitted.values}{Fitted values of the selected model.}

\item{observed.values}{Country specific observed values.}

\item{model.parts}{Object containing detailed results of the fitted model.}

\item{residuals}{Deviance residuals.}
}
\description{
Fit a DoubleGap model for forecasting life expectancy.
The method combines separate forecasts to obtain joint male and female 
life expectancies that are coherent with a best-practice trend. See the entire 
description and mathematical formulation of the model in 
\href{https://doi.org/10.1016/j.insmatheco.2017.09.011}{Pascariu et. al (2017)}.
}
\examples{
# Fit DG model

exF <- MortalityGaps.data$exF
exM <- MortalityGaps.data$exM

M <- DoubleGap(DF = exF, 
               DM = exM, 
               age = 0, 
               country = "AUS", 
               years = 1950:2014, 
               arima.order = c(2, 1, 1), 
               drift = TRUE, 
               tau = 75, 
               A = 86)
M
ls(M)
summary(M)

# Predict model 
P <- predict(M, last_forecast_year = 2050)
P
plot(P)

}
\references{
Pascariu M.D., Canudas-Romo V. and Vaupel W.J. 2017. 
\href{https://doi.org/10.1016/j.insmatheco.2017.09.011}{
The double-gap life expectancy forecasting model.}
Insurance: Mathematics and Economics Volume 78, January 2018, Pages 339-350.
}
\seealso{
\code{\link{predict.DoubleGap}}
}
\author{
Marius D. Pascariu
}
