% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{find.record.ex}
\alias{find.record.ex}
\title{Find record life expectancy at age x given a set of life tables}
\usage{
find.record.ex(data)
}
\description{
Find record life expectancy at age x given a set of life tables
}
\keyword{internal}
