% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{find_fitted_values}
\alias{find_fitted_values}
\title{Find the fitted values of the DoubleGap model}
\usage{
find_fitted_values(M1, M2, M3)
}
\description{
Find the fitted values of the DoubleGap model
}
\keyword{internal}
