% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{find_observed_values}
\alias{find_observed_values}
\title{Find the observed values (from input data). Format.}
\usage{
find_observed_values(x)
}
\description{
Find the observed values (from input data). Format.
}
\keyword{internal}
