% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{find_tau}
\alias{find_tau}
\title{Function to find the value of tau}
\usage{
find_tau(data, f_ = 1.05)
}
\description{
Function to find the value of female life expectancy where the sex-gap is no 
longer expanding (tau). This method is slightly different than the one indicated 
in the original paper (likelihood). The main advantage: speed.
}
\keyword{internal}
