% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hmdlt}
\alias{hmdlt}
\title{Data - for testing purposes}
\format{An object of class \code{list} of length 2.}
\source{
Human Mortality Database, \url{http://www.mortality.org}.
}
\usage{
hmdlt
}
\description{
Dataset containing the HMD life tables.
This data is provided for testing purposes only.
Download the actual data free of charge from \url{http://www.mortality.org}.
}
\keyword{datasets}
