% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityGaps-data.R
\docType{data}
\name{MortalityGaps.data}
\alias{MortalityGaps.data}
\title{DATA - for testing purposes}
\format{
An object of class \code{MortalityGaps.data} of length 2.
}
\source{
\href{https://www.mortality.org}{Human Mortality Database}
}
\usage{
MortalityGaps.data
}
\description{
Dataset containing records of life expectancy at birth and at age 65 for 
female and male populations living in 38 countries between 1950 and 2014.
This dataset is used in \insertCite{pascariu2018;textual}{MortalityGaps} 
article. The data is provided in the package for testing purposes and to 
ensure the reproducibility of the results and figures published in the article. 
By the time you are using this package, the data might be outdated. Download actual 
demographic data free of charge from \insertCite{hmd2017;textual}{MortalityGaps}. 
Once a username and a password is created on the 
\href{https://www.mortality.org}{website} the
\href{https://CRAN.R-project.org/package=MortalityLaws}{MortalityLaws}
R package can be used to extract data directly into your R console.
\insertNoCite{MortalityLaws160;textual}{MortalityGaps}
}
\references{
\insertAllCited{}
}
\seealso{
\code{\link[MortalityLaws]{ReadHMD}}
}
\keyword{datasets}
