% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MortalityGaps-package.R
\docType{package}
\name{MortalityGaps}
\alias{MortalityGaps}
\alias{MortalityGaps-package}
\title{MortalityGaps: The Double-Gap Life Expectancy Forecasting Model}
\description{
Life expectancy is highly correlated over time among countries and between males and females. These associations can be used to improve forecasts. Here we have implemented a method for forecasting female life expectancy based on analysis of the gap between female life expectancy in a country compared with the record level of female life expectancy in the world. Second, to forecast male life expectancy, the gap between male life expectancy and female life expectancy in a country is analysed. We named this method the Double-Gap model. For a detailed description of the method see Pascariu et al. (2018). \doi{10.1016/j.insmatheco.2017.09.011}.
}
\details{
To learn more about the package, start with the vignettes:
\code{browseVignettes(package = "MortalityGaps")}
\insertNoCite{*}{MortalityGaps}
}
\references{
\insertAllCited{}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mpascariu/MortalityGaps}
  \item Report bugs at \url{https://github.com/mpascariu/MortalityGaps/issues}
}

}
\author{
\strong{Maintainer}: Marius D. Pascariu \email{rpascariu@outlook.com} (\href{https://orcid.org/0000-0002-2568-6489}{ORCID})

}
