% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{onAttach}
\alias{onAttach}
\alias{.onAttach}
\title{onAttach}
\usage{
.onAttach(lib, pkg)
}
\arguments{
\item{lib}{lib}

\item{pkg}{pkg}
}
\description{
Print a message when the package is loaded first time in R
}
\keyword{internal}
