% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_predict.R
\name{predict.DoubleGap}
\alias{predict.DoubleGap}
\title{Generic Predict Function for Class \code{\link{DoubleGap}}}
\usage{
\method{predict}{DoubleGap}(object, last_forecast_year, iter = 500,
  ci = c(0.8, 0.95), ...)
}
\arguments{
\item{object}{A DoubleGap object}

\item{last_forecast_year}{Last year to be included in the forecast}

\item{iter}{Number of iterations. Default: 500}

\item{ci}{Confidence levels. Default: c(0.8, 0.95)}

\item{...}{Additional arguments affecting the predictions produced.}
}
\value{
A list containing predicted value of best-practice
life expectancy, best-practice gap, sex gap, and forecast life expectancy 
for females and males (together with prediction intervals).
}
\description{
Predict DoubleGap model
}
\seealso{
\code{\link{DoubleGap}}
}
\author{
Marius D. Pascariu
}
