% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{prepare_data}
\alias{prepare_data}
\title{The role of this function is to prepare data in such a way that is ready to use
right away in the other functions.}
\usage{
prepare_data(data)
}
\description{
The role of this function is to prepare data in such a way that is ready to use
right away in the other functions.
}
\keyword{internal}
