% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{print.DoubleGap}
\alias{print.DoubleGap}
\title{Print DoubleGap}
\usage{
\method{print}{DoubleGap}(x, ...)
}
\arguments{
\item{x}{Object of class \code{DoubleGap}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print DoubleGap
}
\keyword{internal}
