% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_predict.R
\name{print.predict.DoubleGap}
\alias{print.predict.DoubleGap}
\title{Print function for predict.DoubleGap}
\usage{
\method{print}{predict.DoubleGap}(x, ...)
}
\arguments{
\item{x}{An object of class \code{predict.DoubleGap}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print function for predict.DoubleGap
}
\keyword{internal}
