% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{print.summary.DoubleGap}
\alias{print.summary.DoubleGap}
\title{Print summary}
\usage{
\method{print}{summary.DoubleGap}(x, ...)
}
\arguments{
\item{x}{Object of class \code{DoubleGap}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print summary
}
\keyword{internal}
