% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{sex_gap.model}
\alias{sex_gap.model}
\title{Fit sex-gap model. This is a modified version of the Raftery linear model
for sex differences in life expectancy. This is kind of a AR(2)-X
sex_gap ~ sex_gap1 + sex_gap2 + narrow_level}
\usage{
sex_gap.model(data, tau, A)
}
\description{
Fit sex-gap model. This is a modified version of the Raftery linear model
for sex differences in life expectancy. This is kind of a AR(2)-X
sex_gap ~ sex_gap1 + sex_gap2 + narrow_level
}
\keyword{internal}
