% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DoubleGap_S3.R
\name{sim_exf}
\alias{sim_exf}
\title{Simulate female life expectancy (to be used in predict function)}
\usage{
sim_exf(m1, m2, h_, forecast_index, iter)
}
\description{
Simulate female life expectancy (to be used in predict function)
}
\keyword{internal}
