% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{summary.DoubleGap}
\alias{summary.DoubleGap}
\title{Summary DoubleGap}
\usage{
\method{summary}{DoubleGap}(object, ...)
}
\arguments{
\item{object}{Object of class \code{DoubleGap}}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary DoubleGap
}
\keyword{internal}
