% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{bp_gap.model}
\alias{bp_gap.model}
\title{Time series model for the best-practice gap}
\usage{
bp_gap.model(X, benchmark, arima.order, drift)
}
\arguments{
\item{X}{Input data object generated by prepare_data function.}

\item{benchmark}{Fitted values given by BP linear model.}

\item{arima.order}{A specification of the the ARIMA model to be used in 
fitting the best-practice gap. The ARIMA order is country specific.
The three integer components (p, d, q) are the AR order, 
the degree of differencing, and the MA order. Format: numerical vector of length 3.
If \code{arima.order = NULL} the function conducts a search over possible models 
according to AIC. See \code{\link{auto.arima}} for details.}

\item{drift}{Indicate whether the ARIMA model should include a linear drift 
term or not. Type: logical value. If \code{drift = NULL}, it will be estimate 
automatically.}
}
\description{
Time series model for the best-practice gap
}
\keyword{internal}
