% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fun_DG_fit.R
\name{DoubleGap}
\alias{DoubleGap}
\title{Fit DoubleGap model}
\usage{
DoubleGap(LT_female, LT_male, age = 0, country, years, arima.order = c(0, 1,
  0), drift = FALSE, tau = NULL, A = NULL)
}
\arguments{
\item{LT_female}{Life tables for females}

\item{LT_male}{Life tables for males}

\item{age}{Age. The life expectancy model will be fitted at the indicated age.}

\item{country}{Country. Type: character}

\item{years}{Period of time to be used. Type: numeric vector.}

\item{arima.order}{A specification of the the ARIMA model to be used in 
fitting the best-practice gap. The ARIMA order is country specific.
The three integer components (p, d, q) are the AR order, 
the degree of differencing, and the MA order.}

\item{drift}{Indicate whether the ARIMA model should include a linear drift 
term or not. Type: logical value.}

\item{tau}{The level of female life expectancy at which 
the sex-gap is expected to stop widening and to start narrowing. If \code{NULL} 
then the model will run an algorithm to find it.}

\item{A}{The level of female life expectancy where we assume no further 
change in the sex-gap.}
}
\value{
A \code{DoubleGap} object
}
\description{
Fit DoubleGap model
}
\examples{
# Not yet.
}
\author{
Marius Pascariu
}
