% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{DoubleGap}
\alias{DoubleGap}
\title{Fit the DoubleGap Model}
\usage{
DoubleGap(DF, DM, age = 0, country, years, arima.order = c(0, 1, 0),
  drift = FALSE, tau = NULL, A = NULL)
}
\arguments{
\item{DF}{data.frame containing life expectanacy records for females. 
The table must contain the following 4 colums: country, year, age, ex.}

\item{DM}{data.frame containing life expectanacy records for males. 
The table must have the same format and dimensions as \code{DF}.}

\item{age}{Age. The life expectancy model will be fitted at the indicated age.}

\item{country}{Country. Type: character}

\item{years}{Period of time to be used. Type: numeric vector.}

\item{arima.order}{A specification of the the ARIMA model to be used in 
fitting the best-practice gap. The ARIMA order is country specific.
The three integer components (p, d, q) are the AR order, 
the degree of differencing, and the MA order.}

\item{drift}{Indicate whether the ARIMA model should include a linear drift 
term or not. Type: logical value.}

\item{tau}{The level of female life expectancy at which 
the sex-gap is expected to stop widening and to start narrowing. If \code{NULL} 
then the model will run an algorithm to find it.}

\item{A}{The level of female life expectancy where we assume no further 
change in the sex-gap.}
}
\value{
A \code{DoubleGap} object
}
\description{
Fit DoubleGap model
}
\examples{
# Fit DG model

exF <- MortalityGaps.data$exF
exM <- MortalityGaps.data$exM

M <- DoubleGap(DF = exF, 
               DM = exM, 
               age = 0, 
               country = "AUS", 
               years = 1950:2014, 
               arima.order = c(2, 1, 1), 
               drift = TRUE, 
               tau = 75, 
               A = 86)
M
ls(M)
summary(M)

# Predict model 
P <- predict(M, last_forecast_year = 2050)
P
plot(P)

}
\seealso{
\code{\link{predict.DoubleGap}}
}
\author{
Marius D. Pascariu
}
