% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{find_fitted_values}
\alias{find_fitted_values}
\title{Find the fitted values of the DoubleGap model}
\usage{
find_fitted_values(M1, M2, M3)
}
\arguments{
\item{M1}{Model 1 - Linear regression bp-life-expectancy.}

\item{M2}{Model 2 - Time-series D-gap.}

\item{M3}{Model 3 - Sex-gap.}
}
\description{
Find the fitted values of the DoubleGap model
}
\keyword{internal}
