% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{find_observed_values}
\alias{find_observed_values}
\title{Find the observed values (from input data). Format.}
\usage{
find_observed_values(X)
}
\arguments{
\item{X}{Input data object generated by prepare_data function.}
}
\description{
Find the observed values (from input data). Format.
}
\keyword{internal}
