% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{find_record_ex}
\alias{find_record_ex}
\title{Find record life expectancy at age x given a set of life tables}
\usage{
find_record_ex(X)
}
\arguments{
\item{X}{Data-set containing ex records for female and male populations.}
}
\description{
Find record life expectancy at age x given a set of life tables
}
\keyword{internal}
