% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{find_tau}
\alias{find_tau}
\title{Function to find the value of tau}
\usage{
find_tau(X, a = 1.05, f = 0.5)
}
\arguments{
\item{X}{Input data object generated by prepare_data function.}

\item{a}{Adjustment factor for maximum female life expectancy.}

\item{f}{the smoother span. This gives the proportion of points in
    the plot which influence the smooth at each value.
    Larger values give more smoothness.}
}
\description{
Function to find the value of female life expectancy where the sex-gap is no 
longer expanding (tau). This method is slightly different than the one indicated 
in the original paper (likelihood). The main advantage: speed.
}
\keyword{internal}
