% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_predict.R
\name{prepare_data_for_prediction}
\alias{prepare_data_for_prediction}
\title{Prepare additional data in order to perform predictions}
\usage{
prepare_data_for_prediction(object, last_forecast_year, iter, ci)
}
\description{
Prepare additional data in order to perform predictions
}
\keyword{internal}
