% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DGM_fit.R
\name{sex_gap.model}
\alias{sex_gap.model}
\title{Fit sex-gap model. This is a modified version of the Raftery linear model
for sex differences in life expectancy. This is kind of a AR(2)-X
sex_gap ~ sex_gap1 + sex_gap2 + narrow_level}
\usage{
sex_gap.model(X, tau, A)
}
\arguments{
\item{X}{Input data object generated by prepare_data function.}

\item{tau}{The level of female life expectancy at which the sex-gap is 
expected to stop widening and to start narrowing. If \code{NULL} 
then the model will run an algorithm to find it.}

\item{A}{The level of female life expectancy where we assume no further 
change in the sex-gap. If \code{NULL} the model will estimate it.}
}
\description{
Fit sex-gap model. This is a modified version of the Raftery linear model
for sex differences in life expectancy. This is kind of a AR(2)-X
sex_gap ~ sex_gap1 + sex_gap2 + narrow_level
}
\keyword{internal}
